Make a temporary directory for code checking.
This takes the existing path, like /home/user/zones/repo, and builds oun
a new directory name like /home/user/zones/repo_tmp/repo.

It then proceeds to delete anything that happens to already be in that
temporary directory, and then recreates it.

This gives us a clean working slate to examine the new code in.

Note, we should probably try and clean this up at the end of
pre_receive.  Right now the code instead does the clean up at the next
pre_receive, which isn't entirely optimal.