Add check_what_changed to post_receive.
This steals the code that parses stdin and calls check_what_changed from
pre_receive, so that we can do the actual modifications in post_receive.

Now, as we are doing this, the value of the list_file is...  A lot less
than it was, and we could probably get away with just getting rid of it
for pre_receive and post_receive.  At the moment all it does is provide
an optimization for the case where no zones actually change.