Reprocess files in post_receive.
This goes back through and runs process_files, and then does a git
commit if we changed anything.

If pre_receive has already written the changes and committed them, then
this is mostly a no-op, the AUTO_INCREMENT values will be unchanged.

But in the next few commits pre_receive will stop actually making the
changes, and so this will be necessary to keep AUTO_INCREMENT working.