Stop using 'git checkout' in pre_receive.
As of earlier this year, git no longer allows any ref updates in the
pre-receive hook.

This is intended to prevent corruption of your git tree, but it means
that we can't actually do a git checkout in pre-receive.

(See https://www.spinics.net/lists/git/msg300763.html for the git side
details.)

Instead, we are using git archive to dump the tree as of the new commit,
and we are decompressing that in our new temporary directory
(commit 6bbfc5fbe57f1b5788c54ffa5c29b6a535444533).

We then chdir to the temporary directory, do all of our checks, and
switch back to the repo directory before saving the list file.