Follow both $INCLUDE and ;INCLUDED_BY.
With this change, we will attempt to follow changes in both directions,
instead of just looking at INCLUDED_BY.

This means that it is now practical to have your AUTO_INCREMENT serial
number in a file that you $INCLUDE, and it will get updated when you
update any file that includes it.

This can lead to additional unnecessary zone updates, but it may make it
easier to manage a zone which is largely common, but with minor
differences for different views.

If you don't have AUTO_INCREMENT in a file that you $INCLUDE, there
should be no real impact.