Drop the gitzone list for pre/post receive.
post-receive is already doing all of the work needed to rebuild the
list, getting it from pre-receive adds...  Pretty much nothing.

One side effect of this change is that commits that do not alter a
configured zone file now still cause a deployment into $zone_dir.

Now, the question is, what kind of commits can exist, but which don't
touch actual zone files?

I see three possible cases here: Zone files that are not configured yet.
Documentation files.  And people that chose to put named config files in
git along with the zone files.

The first one should be harmless, the second two actually give a benefit
to being deployed.